+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 
+* Enable MIME type detection if libmagic is not available
 * Adds handling for empty files in `detect_filetype` and `partition`.
 
 ### Features